// src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export namespace Models {
  export type public_Lead = {
    aiErrorMessage: CodecTypes['pg/text@1']['output'] | null;
    aiIntent: 'SALES' | 'UNKNOWN' | null;
    aiPriority: 'LOW' | 'MEDIUM' | 'HIGH' | null;
    aiStatus: 'PENDING' | 'COMPLETED' | 'FAILED' | 'FALLBACK';
    aiSummary: CodecTypes['pg/text@1']['output'] | null;
    createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    email: CodecTypes['pg/text@1']['output'];
    id: CodecTypes['pg/text@1']['output'];
    message: CodecTypes['pg/text@1']['output'];
    name: CodecTypes['pg/text@1']['output'];
    source: 'LANDING_PAGE';
    updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    readonly [RelationKeys]?: never;
  };
}